Create Team Management Program: Version 2
Update the team tracking program used in project #1. This program will use a list of dictionary objects where each player is represented by a dictionary of the player’s ‘name’, ‘position’ (POS), ‘at_bats’ (AB),and ‘hits’ (H).  
The programs will consist of one Python file to hold the main method and supporting functions, and a second Python file (FileIO.py) with two functions which read from and write to a comma delimited csv file. The program will use a file named players.csv to store the data. Remember that each row in a CSV reader object is a list and each column in the row can be accessed with a list index. 
The data is loaded into a list in the main method before the user is asked for a menu option. The list will be used as an argument when calling a support function. 
The separator lines are 60 characters based on the multiplication operator. The format for the lineup display columns will use 3 spaces for the order number, 31 spaces for the player name, 6 spaces for the position, at bats, and hits, and 8 spaces for the batting average. The batting average must always have 3 decimals.
Use a tuple to represent the valid positions a player can have.